Add some basic security checks
remote-method-guessing was vulnerable to shell injection attacks. When
scanning a malicious RMI server, the server could potentially inject
shell and malicious java commands on the executing hosts.

This issue was caused as 'bound names' and 'class names' were used
unfiltered inside the code that is generated by remote-method-guesser.
Users can identify potentially malicious servers by first running rmg
without the --guess argument and inspect the bound and class names. But
if --guess is used directly, it can lead to dangerous results.
(Notice that methods are no potential injection target, as method names
and arguments are defined as part of the template files).

With this commit, some basic filtering for malicious bound and class
names was added. This filtering was chosen quite restrictive so that rmg
does probably no longer work on certain exoctic class names. In future
commits, a --trsuted option is added to allow disabling these filters.