Update template creation process
Where was a general error when building sample files.

When building a sample file, the contents of the corresponding interface
were included into the sample. This worked fine for the test-docker
container, as the sample as well as the RMI server classes are part of
the same package (de.qtc.rmg). However, once you have classes from a
different package, the interface from within the sample is no longer
valid.

To solve this issue, the compiled interface is now copied during the
sample creation into the sample folder. Additionally, the sample package
declaration was removed.