Add first implementaiton of type confusing
Added a first implementation of type confusing. The paramaterTypes class
array of remote methods is now modified before the invocation. This will
still cause fingerprintable error messages on the server side, but at
the same time the remote method is never really invoked.

This funcationality was inspired by the rmiscout project, although it is
implemented slightly different in the remote-method-gusser.