Use lookupCall for default RMI bypass
The bypass action now uses the rmiregistry.lookup call per default for
the bypass technique. This bypasses the hostname restriction that is
normally applied to the bind, unbind and rebind calls.

As the serialization filter bypass was patched to the same time as the
'string-restriction' was applied for deserilaization on RMI endpoints,
it should be no restriction. Furthermore, you can use the --local option
to use the rmiregistry.bind call instead.