Remove --json option
JSON output was only barely implemented for bound names and successfully
guessed methods. It is not implemented for all the new stuff that has
been implemented. All the JSON related functions are still present
within the code, but ``--json`` was removed from the argument parser.
We decide in future whether to add JSON support back for all functions
or to remove it completly.