Add String unmarshalling enumeration
The enum action of rmg now performs a check how String is unmarshalled by
the remote server. This is done by calling the registry method lookup
with custom objects. As the lookup method expects a String, exceptions
will always be thrown. However, depending on the exception type, it is
possible to determine whether readObject or readString was used to
umarshall the input argument.