[docker] Use customized jre (jlink)
The container now uses a customized jre that was generated by jlink.
This reduces the overall container size down to ~35MB. Furthermore, the
container is now available in two different versions (Java9 & Java11).
These versions were chosen because:

1. They support jlink
2. Java9 uses readObject, whereas Java11 uses readString for
   unmarshalling the String type during RMI calls.

Notice that the example-server still needs to be build on Java8, as
statically compiled stubs are no longer supported in other releases
(there are probably workarounds, but it should not matter).