The grand refactoring
This way to large commit refactores many functions that were previously
contained in the main function. Furthermore, it introfuces a first
version of a full functioning plugin system and resolves issues during
marshalling when using custom RMI calls.

All issues are somehow connected and are therefore commit as one large
change. This probably contains still a lot of bugs that will be resolved
in future smaller commits.