Fix internal deserialization trigger
During one of the latest changes on the development branch, we changed
the marshalling behavior of the raw RMI call function and enforced the
usage of object <-> type pairs for RMI calls.

The corresponding pairs were stored within a LinkedHashMap, that caused
issues when performing deserialization attacks. As some gadgets trigger
when put into a HashMap, it was possible that the corresponding payload
was executed locally instead on the target.

This is obviously pretty bad, but was luckily never pushed to the master
branch. With this commit, the behavior is fixed. Payload objects are now
stored within a dedicated data type that does not trigger magic methods
to be executed.