Change String marshaling detection method
The string marshaling detection was previously based on sending a
malformed URL within the object annotation and looking for an
MalformedUrlException. However, this only works for servers that have
useCodebaseOnly=false set. The enumeration succeeded anyway, as there
was a fallback implemented that compares exception messages against
readString to evaluate the String marshaling.

With this commit, String marshaling is enumerated by sending an unknown
class within the object annotation. This should always lead to a
ClassNotFound exception, independent of the value of useCodebaseOnly.