Store wordlists and template within of jar
Before this commit, rmg did only support external wordlist and template
files that need to be available on the file system. The downside of this
is that you either need to specify the path manually or configure a
default path that needs to be modified by the user.

Now rmg includes it's default wordlist and templates within of the .jar
file, which makes it more flexible.