Make logHit atomic
Not really, but I remeber that we had some problems in the early stages
of the project when we used multiple print calls to log a result. The
print calls mixed up in this case and the solution was to use a single
print call instead. Not sure whether this is really how one should do
it, but it seems to work fine :D