Catch UnknownHostException in SockFactory classes
A possibly occuring UnknownHostException was catched within the
corresponding RMI call function so far. The problem with this approach
is that the target of the RMI call may differs from the specified target
on the command line (when using the --follow option).

To get the real target, the RMI call functions would need to obtain the
TCPEndpoint from the LiveRef, that is encapsulated in the UnicastRef.
This would require some more reflection and is probably not the cleanest
approach.

Since we already overwrite the socket factories that are used to
establish connections, we catch the UnknownHostException now directly at
this point.