Introduce ScanActions
This commit introduces ScanActions, a feature that can be used to select
which operations are exectuted within the rmg enum action. The desired
scan actions are just specified after the enum keyword on the command
line.

With this implementation it should be possible to perform any rmg action
now also via SSRF.