Allow multiple class names for KnownEndpoint
Before this commit, a KnownEndpoint was bound to a single class name.
This can be problematic for legacy remote objects that can be exported
as stubs as well as proxy objects. Therefore, the possibility to assign
mutliple classNames to a KnownEndpont was now added.

Also added comments for all new implemented classes and functions.