Fix bug in getThrowable and fix BindExceptions
The getThrowable method (searches for a specific throwable within a
stack trace) did not work when the first throwable was already matching.
Now this was changed and the first throwable is returned if it is
matching.

Based on that, we fixed the handling of BindExceptions, which caused
malformed exception messages in some cases.