Change enumDGC to enumSecurityManager
The enumDGC function made some incorrect assumptions on the separate
AccessControlContext used by the DGC. The function assumed that it is
possible to detect outdated DGC implementations by inspecting the
exception messages. However, this was not correct and the results were
meaningless. Therefore, we switched the function to enumSecurityManager,
because this is the only thing that can be enumerated reliably.