Add roguejmx action
Added the roguejmx actions, which creates a rogue JMX service. This
service acts like an ordinary JMX service, but captures credentials of
incoming connections. Connections can optionally be forwarded to a real
JMX service.

This action was mainly implemented to demonstrate the impact of registry
modifications made by a malicious client.