Make RogueJMX Java8 comptaible
When implementing the RogueJMX server, the compatibility to Java 8 got
broken as we used java.io.ObjectInputFilter. This class only exists for
Java 9+ and is represented by sun.misc.ObjectInputFilter in Java 8 and
earlier.

To solve this issue, we now use reflective access to the
ObjectInputFilter class and inject the filter into UnicastServerRef
manually using reflection. It is not beautiful, but it seems to work :)