Improve completion script
The completion script had problems when value options were used. Some
completions used cword to determine what to complete. However, cword
does not exclude value options and is therefore not suitable for this
case. Instead, $args is the correct choice, as it excludes value options
and counts from start of the command line to the current word.