[docker] Add activation system to example server
Added an activation system to the example server. This can be used to
reproduce the error mentioned in issue #30. The code still throws one
exception where I'm not sure from where its coming from. However, the
server seems to be functional.

This commit probably crashes the pipeline, as the activation system was
removed from more recent jdk packages. We will need to investigate whats
the best way to build it in future.