Add first draft for activation support
Added a first draft for activation support that addresses issue #30.
The basic operations like enum should already be fully functional for
java versions that still containt the activation system related classes.
More complex operations need to be tested and may not be fully
functional yet for ActivatableRef types. Furthermore, support for java
versions that miss the activation classes needs to be implemented.