Add dynamic support for uncommon serialVersionUIDs
As outlined in issue #39, legacy RMI stubs may be compiled with an
uncommon serialVersionUID that does not match the default value 2L. The
latest commits on the development branch addressed this issue by letting
the user select a serialVersionUID using the --serial-version-uid
option. While this worked wirh a single bound name using an uncommon
serialVersionUID, it did not work with multiple one.

This commit implements another solution for this issue. When an uncommon
serialVersionUID is encountered, the observed serialVersionUID is saved
with the associated class name and the corresponding boundname is resolved
again. Within the RMI classloader we check for each class whether a non
default serialVersionUID was observed. If this was the case, we recreate
the corresponding class with the correct serialVersionUID. Since the
class was already created during the first lookup, the fully qualified
name of the new class is prefixed with an underscore.