Improve reliability of --serial-version-uid
When using the --serial-version-uid option, the lookup operation
sometimes causes an invalid transport code exeception. This one seems to
be some kind of race condition and cannot be triggered reliably. Since
it is often sufficient to retry the operation, rmg now automatically
retries up to five times after encountering this exception.