Increase max lookup-loop count
When remote-method-guesser performs a lookup operation, classes may be
created dynamically to handle custom socket factories or RMI stubs. The
serialVersionUid of these classes need to match the incomming class.
remote-method-guesser hadles this using an error based appraoch, where
the required serialVersionUid is obtained from an exception message.

Up to this commit, the above mentioned mechanism only triggered once.
This causes problems if you have e.g. a RMI stub class that uses a
custom socket factory too. With this commit, the count is increased to
five (probably more than required, but it won't hurt).