Improve exception handling during method guessing
In some configurations, stream corruptions can occur during method
guessing with multiple threads. This behavior seems to be non
deterministically and was not fully understood yet. However, even on
affected configurations, the miss rate is roughly at 1/3000 methods that
fail. For this reason, we catch the resulting exceptions for now and
only display warnings if running in verbose mode.